+++ bundle/src/main/java/com/adobe/acs/commons/contentsync/RemoteInstance.java
@@ -20,6 +20,7 @@
 package com.adobe.acs.commons.contentsync;
 
 import com.adobe.acs.commons.adobeio.service.IntegrationService;
+import org.apache.http.HttpHost;
 import org.apache.http.HttpRequestInterceptor;
 import org.apache.http.HttpStatus;
 import org.apache.http.auth.AuthScope;
@@ -35,74 +36,77 @@ import org.apache.http.conn.ssl.TrustSelfSignedStrategy;
 import org.apache.http.impl.client.BasicCredentialsProvider;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClientBuilder;
-import org.apache.http.impl.client.HttpClients;
+import org.apache.http.osgi.services.HttpClientBuilderFactory;
 import org.apache.http.ssl.SSLContextBuilder;
 import org.apache.http.util.EntityUtils;
 import org.apache.sling.api.resource.ValueMap;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import javax.json.Json;
 import javax.json.JsonObject;
 import javax.json.JsonReader;
 import javax.json.JsonValue;
 import java.io.Closeable;
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
+import java.lang.invoke.MethodHandles;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.security.GeneralSecurityException;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 import java.util.stream.Collectors;
 
-import static com.adobe.acs.commons.contentsync.ConfigurationUtils.CONNECT_TIMEOUT_KEY;
-import static com.adobe.acs.commons.contentsync.ConfigurationUtils.DISABLE_CERT_CHECK_KEY;
-import static com.adobe.acs.commons.contentsync.ConfigurationUtils.SO_TIMEOUT_STRATEGY_KEY;
 import static org.apache.jackrabbit.JcrConstants.JCR_PRIMARYTYPE;
 
 /**
  * HTTP connection to a remote AEM instance + some sugar methods to fetch data
  */
 public class RemoteInstance implements Closeable {
-    static final int CONNECT_TIMEOUT = 5000;
-    static final int SOCKET_TIMEOUT = 300000;
+    private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 
     private final CloseableHttpClient httpClient;
     private final SyncHostConfiguration hostConfiguration;
 
-    /**
-     *
-     * @deprecated
-     */
     @Deprecated
     public RemoteInstance(SyncHostConfiguration hostConfiguration, ValueMap generalSettings) throws GeneralSecurityException, IOException {
-        this(hostConfiguration, generalSettings, null);
+        throw new RuntimeException("@Deprecated");
     }
 
+    @Deprecated
     public RemoteInstance(SyncHostConfiguration hostConfiguration, ValueMap generalSettings, IntegrationService integrationService) throws GeneralSecurityException, IOException {
-        this.hostConfiguration = hostConfiguration;
-        this.httpClient = createHttpClient(hostConfiguration, generalSettings,  integrationService);
+        throw new RuntimeException("@Deprecated");
     }
 
-    private CloseableHttpClient createHttpClient(SyncHostConfiguration hostConfiguration, ValueMap generalSettings, IntegrationService integrationService)
+    public RemoteInstance(HttpClientBuilderFactory builderFactory, SyncHostConfiguration hostConfiguration, GeneralSettingsModel generalSettings, IntegrationService integrationService) throws GeneralSecurityException, IOException {
+        this.hostConfiguration = hostConfiguration;
+        this.httpClient = createHttpClient(builderFactory.newBuilder(), hostConfiguration, generalSettings,  integrationService);
+    }
+
+    private CloseableHttpClient createHttpClient(HttpClientBuilder builder, SyncHostConfiguration hostConfiguration, GeneralSettingsModel generalSettings, IntegrationService integrationService)
             throws GeneralSecurityException {
-        HttpClientBuilder builder = HttpClients.custom();
         setAuthentication(hostConfiguration, builder, integrationService);
-        int soTimeout = generalSettings.get(SO_TIMEOUT_STRATEGY_KEY, SOCKET_TIMEOUT);
-        int connTimeout = generalSettings.get(CONNECT_TIMEOUT_KEY, CONNECT_TIMEOUT);
-        boolean disableCertCheck = generalSettings.get(DISABLE_CERT_CHECK_KEY, false);
-        RequestConfig requestConfig = RequestConfig
+        int soTimeout = generalSettings.getSocketTimeout();
+        int connTimeout = generalSettings.getConnectTimeout();
+        boolean disableCertCheck = generalSettings.isDisableCertCheck();
+        RequestConfig.Builder requestConfig = RequestConfig
                 .custom()
                 .setConnectTimeout(connTimeout)
                 .setSocketTimeout(soTimeout)
-                .setCookieSpec(CookieSpecs.STANDARD).build();
-        builder.setDefaultRequestConfig(requestConfig);
+                .setCookieSpec(CookieSpecs.STANDARD);
+        String proxyHost = System.getenv("AEM_PROXY_HOST");
+        if (proxyHost != null) {
+            int proxyPort = Integer.parseInt(System.getenv().getOrDefault("AEM_HTTPS_PROXY_PORT", "3128"));
+            log.debug("AEM_PROXY_HOST: {}, AEM_HTTPS_PROXY_PORT: {}", proxyHost, proxyPort);
+            requestConfig.setProxy(new HttpHost(proxyHost, proxyPort));
+        }
+        builder.setDefaultRequestConfig(requestConfig.build());
         if (disableCertCheck) {
             // Disable hostname verification and allow self-signed certificates
             SSLContextBuilder sslbuilder = new SSLContextBuilder();
             sslbuilder.loadTrustMaterial(new TrustSelfSignedStrategy());
-            SSLConnectionSocketFactory sslsf = null;
-            sslsf = new SSLConnectionSocketFactory(
+            SSLConnectionSocketFactory sslsf = new SSLConnectionSocketFactory(
                     sslbuilder.build(), NoopHostnameVerifier.INSTANCE);
             builder.setSSLSocketFactory(sslsf);
         }
@@ -112,7 +116,6 @@ public class RemoteInstance implements Closeable {
     void setAuthentication(SyncHostConfiguration hostConfiguration, HttpClientBuilder builder, IntegrationService integrationService){
         if (hostConfiguration.isOAuthEnabled()) {
             // If OAuth is enabled, use the AccessTokenProvider to get the token
-            // The session user must have the private key from the Adobe technical account installed in the user key store
             try {
                 // the lifetime of Adobe's tokens is 24 hours, enough to request once and re-use across all the calls
                 String accessToken = integrationService.getAccessToken();
@@ -145,26 +148,36 @@ public class RemoteInstance implements Closeable {
     }
 
     public InputStream getStream(URI uri ) throws IOException {
+        log.debug("getStream: {}", uri);
         HttpGet request = new HttpGet(uri);
         CloseableHttpResponse response = httpClient.execute(request);
-        String msg;
-        switch (response.getStatusLine().getStatusCode()){
-            case HttpStatus.SC_OK:
-                return response.getEntity().getContent();
-            case HttpStatus.SC_MULTIPLE_CHOICES:
-                msg = formatError(uri.toString(), response.getStatusLine().getStatusCode(),
-                        "It seems that the \"Json Max Results\" in Sling Get Servlet is too low. Increase it to a higher value, e.g. 1000.");
-                throw new IOException(msg);
-            default:
-                msg = formatError(uri.toString(), response.getStatusLine().getStatusCode(), "Response: " + EntityUtils.toString(response.getEntity()));
-                throw new IOException(msg);
+        int statusCode = response.getStatusLine().getStatusCode();
+        log.debug("getStream() statusCode: {}", response.getStatusLine().getStatusCode());
+        if (statusCode == HttpStatus.SC_OK){
+            return response.getEntity().getContent();
+        } else {
+            String textResponse = response.getEntity() == null ? "" : EntityUtils.toString(response.getEntity());
+            String msg;
+            switch (statusCode){
+                case HttpStatus.SC_MULTIPLE_CHOICES:
+                    msg = formatError(uri.toString(), response.getStatusLine().getStatusCode(),
+                            "It seems that the \"Json Max Results\" in Sling Get Servlet is too low. Increase it to a higher value, e.g. 1000.");
+                    throw new IOException(msg);
+                case HttpStatus.SC_NOT_FOUND:
+                    throw new FileNotFoundException("Not found: " + uri);
+                default:
+                    msg = formatError(uri.toString(), response.getStatusLine().getStatusCode(), "Response: " + textResponse);
+                    throw new IOException(msg);
+            }
         }
     }
 
     public String getString(URI uri) throws IOException {
+        log.debug("getString: {}", uri);
         HttpGet request = new HttpGet(uri);
         try (CloseableHttpResponse response = httpClient.execute(request)) {
             String str = EntityUtils.toString(response.getEntity());
+            log.debug("getString() statusCode: {}", response.getStatusLine().getStatusCode());
             if (response.getStatusLine().getStatusCode() == HttpStatus.SC_OK) {
                 return str;
             } else {
@@ -190,6 +203,7 @@ public class RemoteInstance implements Closeable {
 
     public List<String> listChildren(String path) throws IOException, URISyntaxException {
         List<String> children;
+        log.debug("listChildren: {}", path);
         try (InputStream is = getStream(path + ".1.json"); JsonReader reader = Json.createReader(is)) {
             children = reader
                     .readObject()
@@ -198,6 +212,8 @@ public class RemoteInstance implements Closeable {
                     .filter(entry -> entry.getValue().getValueType() == JsonValue.ValueType.OBJECT)
                     .map(Map.Entry::getKey)
                     .collect(Collectors.toList());
+        } catch (FileNotFoundException e){
+            children = Collections.emptyList();
         }
         return children;
     }
@@ -230,5 +246,9 @@ public class RemoteInstance implements Closeable {
     @Override
     public void close() throws IOException {
         httpClient.close();
+    }
+
+    public SyncHostConfiguration getHostConfiguration(){
+        return hostConfiguration;
     }
 }