+++ bundle/src/main/java/com/adobe/acs/commons/redirects/models/package-info.java
@@ -15,5 +15,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-@org.osgi.annotation.versioning.Version("6.12.0")
+@org.osgi.annotation.versioning.Version("6.15.0")
 package com.adobe.acs.commons.redirects.models;